test(inbound): test Peers requests to the inbound service, directly and via TCP
