+++ zebrad/src/components/inbound/tests/real_peer_set.rs
@@ -1,0 +1,286 @@
+//! Inbound service tests with a real peer set.
+
+use std::{collections::HashSet, net::SocketAddr, sync::Arc};
+
+use futures::FutureExt;
+use tokio::{sync::oneshot, task::JoinHandle};
+use tower::{
+    buffer::Buffer,
+    builder::ServiceBuilder,
+    util::{BoxCloneService, BoxService},
+    ServiceExt,
+};
+
+use zebra_chain::{
+    block::{self, Block},
+    parameters::Network,
+};
+use zebra_consensus::{chain::VerifyChainError, error::TransactionError, transaction};
+use zebra_network::{connect_isolated_tcp_direct, Config as NetworkConfig, Request, Response};
+use zebra_state::Config as StateConfig;
+use zebra_test::mock_service::{MockService, PanicAssertion};
+
+use crate::{
+    components::{
+        inbound::{Inbound, InboundSetupData},
+        mempool::{self, gossip_mempool_transaction_id, Mempool},
+        sync::{self, BlockGossipError, SyncStatus},
+    },
+    BoxError,
+};
+
+/// Check that a network stack with an empty address book only contains the local listener port,
+/// by querying the inbound service via a local TCP connection.
+///
+/// Uses a real Zebra network stack with a local listener address,
+/// and an isolated Zebra inbound TCP connection.
+#[tokio::test]
+async fn inbound_peers_empty_address_book() -> Result<(), crate::BoxError> {
+    let (
+        // real services
+        connected_peer_service,
+        inbound_service,
+        _peer_set,
+        _mempool_service,
+        _state_service,
+        // mocked services
+        _mock_block_verifier,
+        _mock_tx_verifier,
+        // real tasks
+        block_gossip_task_handle,
+        tx_gossip_task_handle,
+        // real open socket addresses
+        listen_addr,
+    ) = setup().await;
+
+    // Use inbound directly
+    let request = inbound_service.clone().oneshot(Request::Peers);
+    let response = request.await;
+    match response.as_ref() {
+        Ok(Response::Peers(single_peer)) if single_peer.len() == 1 => {
+            assert_eq!(single_peer.first().unwrap().addr(), listen_addr)
+        }
+        Ok(Response::Peers(_peer_list)) => unreachable!(
+            "`Peers` response should contain a single peer, \
+             actual result: {:?}",
+            response
+        ),
+        _ => unreachable!(
+            "`Peers` requests should always respond `Ok(Response::Peers(_))`, \
+             actual result: {:?}",
+            response
+        ),
+    };
+
+    // Use the connected peer via a local TCP connection
+    let request = connected_peer_service.clone().oneshot(Request::Peers);
+    let response = request.await;
+    match response.as_ref() {
+        Ok(Response::Peers(single_peer)) if single_peer.len() == 1 => {
+            assert_eq!(single_peer.first().unwrap().addr(), listen_addr)
+        }
+        Ok(Response::Peers(_peer_list)) => unreachable!(
+            "`Peers` response should contain a single peer, \
+             actual result: {:?}",
+            response
+        ),
+        _ => unreachable!(
+            "`Peers` requests should always respond `Ok(Response::Peers(_))`, \
+             actual result: {:?}",
+            response
+        ),
+    };
+
+    let block_gossip_result = block_gossip_task_handle.now_or_never();
+    assert!(
+        matches!(block_gossip_result, None),
+        "unexpected error or panic in block gossip task: {:?}",
+        block_gossip_result,
+    );
+
+    let tx_gossip_result = tx_gossip_task_handle.now_or_never();
+    assert!(
+        matches!(tx_gossip_result, None),
+        "unexpected error or panic in transaction gossip task: {:?}",
+        tx_gossip_result,
+    );
+
+    Ok(())
+}
+
+/// Setup a real Zebra network stack, with a connected peer using a real isolated network stack.
+///
+/// Uses fake verifiers, and does not run a block syncer task.
+async fn setup() -> (
+    // real services
+    // connected peer
+    Buffer<
+        BoxService<zebra_network::Request, zebra_network::Response, BoxError>,
+        zebra_network::Request,
+    >,
+    // inbound service
+    BoxCloneService<zebra_network::Request, zebra_network::Response, BoxError>,
+    // outbound peer set (only has the connected peer)
+    Buffer<
+        BoxService<zebra_network::Request, zebra_network::Response, BoxError>,
+        zebra_network::Request,
+    >,
+    Buffer<BoxService<mempool::Request, mempool::Response, BoxError>, mempool::Request>,
+    Buffer<BoxService<zebra_state::Request, zebra_state::Response, BoxError>, zebra_state::Request>,
+    // mocked services
+    MockService<Arc<Block>, block::Hash, PanicAssertion, VerifyChainError>,
+    MockService<transaction::Request, transaction::Response, PanicAssertion, TransactionError>,
+    // real tasks
+    JoinHandle<Result<(), BlockGossipError>>,
+    JoinHandle<Result<(), BoxError>>,
+    // real open socket addresses
+    SocketAddr,
+) {
+    zebra_test::init();
+
+    let network = Network::Mainnet;
+    // Open a listener on an unused IPv4 localhost port
+    let config_listen_addr = "127.0.0.1:0".parse().unwrap();
+
+    // Inbound
+    let (setup_tx, setup_rx) = oneshot::channel();
+    let inbound_service = Inbound::new(setup_rx);
+    let inbound_service = ServiceBuilder::new()
+        .boxed_clone()
+        .load_shed()
+        .buffer(10)
+        .service(inbound_service);
+
+    // State
+    let state_config = StateConfig::ephemeral();
+    let (state_service, latest_chain_tip, chain_tip_change) =
+        zebra_state::init(state_config, network);
+    let state_service = ServiceBuilder::new().buffer(10).service(state_service);
+
+    // Network
+    let network_config = NetworkConfig {
+        network,
+        listen_addr: config_listen_addr,
+
+        // Stop Zebra making outbound connections
+        initial_mainnet_peers: HashSet::new(),
+        initial_testnet_peers: HashSet::new(),
+
+        ..NetworkConfig::default()
+    };
+    let (mut peer_set, address_book) = zebra_network::init(
+        network_config,
+        inbound_service.clone(),
+        latest_chain_tip.clone(),
+    )
+    .await;
+
+    // Inbound listener
+    let listen_addr = address_book
+        .lock()
+        .unwrap()
+        .local_listener_meta_addr()
+        .addr();
+    assert_ne!(
+        listen_addr.port(),
+        0,
+        "dynamic ports are replaced with OS-assigned ports"
+    );
+    assert_eq!(
+        listen_addr.ip(),
+        config_listen_addr.ip(),
+        "IP addresses are correctly propagated"
+    );
+
+    // Fake syncer
+    let (sync_status, mut recent_syncs) = SyncStatus::new();
+
+    // Fake verifiers
+    let mock_block_verifier = MockService::build().for_unit_tests();
+    let buffered_block_verifier = ServiceBuilder::new()
+        .buffer(10)
+        .service(BoxService::new(mock_block_verifier.clone()));
+    let mock_tx_verifier = MockService::build().for_unit_tests();
+    let buffered_tx_verifier = ServiceBuilder::new()
+        .buffer(10)
+        .service(BoxService::new(mock_tx_verifier.clone()));
+
+    // Mempool
+    let mempool_config = mempool::Config::default();
+    let (mut mempool_service, transaction_receiver) = Mempool::new(
+        &mempool_config,
+        peer_set.clone(),
+        state_service.clone(),
+        buffered_tx_verifier.clone(),
+        sync_status.clone(),
+        latest_chain_tip.clone(),
+        chain_tip_change.clone(),
+    );
+
+    // Enable the mempool
+    mempool_service.enable(&mut recent_syncs).await;
+    let mempool_service = ServiceBuilder::new()
+        .buffer(10)
+        .boxed()
+        // boxed() needs this extra tiny buffer
+        .buffer(1)
+        .service(mempool_service);
+
+    // Initialize the inbound service
+    let setup_data = InboundSetupData {
+        address_book,
+        block_download_peer_set: peer_set.clone(),
+        block_verifier: buffered_block_verifier,
+        mempool: mempool_service.clone(),
+        state: state_service.clone(),
+        latest_chain_tip,
+    };
+    let r = setup_tx.send(setup_data);
+    // We can't expect or unwrap because the returned Result does not implement Debug
+    assert!(r.is_ok(), "unexpected setup channel send failure");
+
+    let block_gossip_task_handle = tokio::spawn(sync::gossip_best_tip_block_hashes(
+        sync_status.clone(),
+        chain_tip_change,
+        peer_set.clone(),
+    ));
+
+    let tx_gossip_task_handle = tokio::spawn(gossip_mempool_transaction_id(
+        transaction_receiver,
+        peer_set.clone(),
+    ));
+
+    // Open a fake peer connection to the inbound listener, using the isolated connection API
+    let connected_peer_service =
+        connect_isolated_tcp_direct(network, listen_addr, "test".to_string())
+            .await
+            .expect("local listener connection succeeds");
+    let connected_peer_service = ServiceBuilder::new()
+        .buffer(10)
+        .service(connected_peer_service);
+
+    // Make the peer set find the new peer
+    let _ = peer_set
+        .ready()
+        .await
+        .expect("peer set becomes ready without errors");
+
+    // there is no syncer task, and the verifiers are fake,
+    // but the network stack is all real
+    (
+        // real services
+        connected_peer_service,
+        inbound_service,
+        peer_set,
+        mempool_service,
+        state_service,
+        // mocked services
+        mock_block_verifier,
+        mock_tx_verifier,
+        // real tasks
+        block_gossip_task_handle,
+        tx_gossip_task_handle,
+        // real open socket addresses
+        listen_addr,
+    )
+}
